+++ core/import-cobalt.c
@@ -6,6 +6,7 @@
 
 #include "ssrf.h"
 #include "divesite.h"
+#include "gas.h"
 #include "subsurface-string.h"
 #include "parse.h"
 #include "divelist.h"
