+++ core/libdivecomputer.c
@@ -19,6 +19,7 @@
 #include "dive.h"
 #include "display.h"
 #include "errorhelper.h"
+#include "event.h"
 #include "sha1.h"
 #include "subsurface-time.h"
 #include "timer.h"
