+++ core/statistics.c
@@ -12,6 +12,7 @@
 
 #include "dive.h"
 #include "display.h"
+#include "event.h"
 #include "subsurface-time.h"
 #include "trip.h"
 #include "statistics.h"
